more progress on drag and drop
removed geometry reader from ghost item because the translated position
can be derived from the original position of the item dragged, added with
the translation. However, sometimes the position of the ghost item is
off possibly because of not accounting for where you drag from. If the
desired ghost item position while dragging is to be under the finger it
will have to account for the original position from where you started
the drag within the item.